pyproject.toml
[project]
name = "llm-rag"
version = "0.1.0"
description = "LLM RAG"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-chroma>=0.2.2",
    "langchain-community>=0.3.18",
    "langchain-openai>=0.3.7",
    "langgraph>=0.2.74",
    "langgraph-cli[inmem]>=0.1.74",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
]

[dependency-groups]
dev = ["ipywidgets>=8.1.5", "jupyterlab>=4.3.5", "ruff>=0.9.7"]
